feat(verify-prd): add FAIL path (shipped->blocked + failure report + linked fix issues)
Extends the verify-prd skill (#597 scaffold + #598 PASS path/verification
core) with the FAIL branch of PRD-level verification. When spec conformance
returns PARTIAL/DIVERGES (CONFORMANCE_FAILED) or any applicable empirical
check fails (EMPIRICAL_FAILED), the skill now:

- transitions the PRD shipped -> blocked, reusing the existing `blocked`
  role per prd-lifecycle-rollup's "no extra failure states" rule (no new
  prd-verification-failed / prd-verifying state), resolved vendor-neutrally
  via config-resolution (Phase 7.1/7.2);
- posts a product-readable failure report on the PRD naming the specific
  failed requirements/ACs with observed-vs-expected evidence (Phase 7.3);
- creates one or more linked fix issues via tracker-write — each carrying
  Gherkin acceptance criteria and captured evidence, back-linked to the
  PRD and the failure report, referencing the failed requirement, and NOT
  reopening the already-terminal generated children (Phase 7.4).

Idempotency (#600) remains out of scope. Cites prd-lifecycle-rollup by
slug for the shipped -> blocked FAIL hop and config-resolution for the
blocked role vocabulary.

Tests: updates the #598 PASS-path doc-presence test that asserted non-pass
branches stay at shipped (now they run the FAIL path), and adds a #599
FAIL-path block covering all four acceptance scenarios across both plugin
roots. Splits the executable guard-logic proof into a sibling test file to
stay within the max-lines budget.

Source edited under plugins/src; plugins/lisa regenerated via bun run
build:plugins. Closes #599.
